Interface graphique fonctionnelle quelques legers bugs a fix

+++ src/main/java/org/phantastikos/entite/creatures/Elfe.java
@@ -5,7 +5,7 @@ import org.phantastikos.entite.creatures.comportements.Vip;
 
 public class Elfe extends Creature implements Vip {
 
-    public Elfe(String nom, char sexe, double poids, double taille, int age) {
+    public Elfe(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(80);
     }
