+++ src/main/java/org/phantastikos/entite/creatures/HommeBete.java
@@ -4,7 +4,7 @@ import org.phantastikos.entite.creatures.comportements.Triage;
 
 public class HommeBete extends Creature implements Triage {
 
-    public HommeBete(String nom, char sexe, double poids, double taille, int age) {
+    public HommeBete(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(80);
     }
