+++ src/main/java/org/phantastikos/entite/creatures/Nain.java
@@ -4,7 +4,7 @@ import org.phantastikos.entite.creatures.comportements.Vip;
 
 public class Nain extends Creature implements Vip {
 
-    public Nain(String nom, char sexe, double poids, double taille, int age) {
+    public Nain(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(95);
     }
