+++ src/main/java/org/phantastikos/entite/creatures/Orque.java
@@ -4,7 +4,7 @@ import org.phantastikos.entite.creatures.comportements.Triage;
 
 public class Orque extends Creature implements Triage {
 
-    public Orque(String nom, char sexe, double poids, double taille, int age) {
+    public Orque(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(60);
     }
