+++ src/main/java/org/phantastikos/entite/creatures/Reptilien.java
@@ -4,7 +4,7 @@ import org.phantastikos.entite.creatures.comportements.Vip;
 
 public class Reptilien extends Creature implements Vip {
 
-    public Reptilien(String nom, char sexe, double poids, double taille, int age) {
+    public Reptilien(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(80);
     }
