+++ src/main/java/org/phantastikos/entite/creatures/Vampire.java
@@ -8,7 +8,7 @@ import org.phantastikos.entite.creatures.comportements.Vip;
 
 public class Vampire extends Creature implements Vip, Demoralisant, Contaminant, Regenerant {
 
-    public Vampire(String nom, char sexe, double poids, double taille, int age) {
+    public Vampire(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(100);
     }
