+++ src/main/java/org/phantastikos/entite/creatures/Zombie.java
@@ -5,7 +5,7 @@ import org.phantastikos.entite.creatures.comportements.Triage;
 
 public class Zombie extends Creature implements Triage {
 
-    public Zombie(String nom, char sexe, double poids, double taille, int age) {
+    public Zombie(String nom, char sexe, int poids, int taille, int age) {
         super(nom, sexe, poids, taille, age);
         setMoral(10);
     }
